fix: change author info placement
