+++ src/assets/authorInfo.ts
@@ -1,0 +1,7 @@
+export const gitHubInfo = {
+  href: 'https://github.com/YuliAdam',
+  text: '@YuliAdam',
+  year: '2025',
+  schoolName: 'RSchool',
+  courseLink: 'https://rs.school/courses/reactjs',
+};
